requirements.txt
face_recognition
numpy
Pillow
ipython